requirements.txt
pandas
numpy
matplotlib
seaborn
openpyxl
pillow
sys
os
pathlib

